backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Scraping
scrapling==0.2.5
feedparser==6.0.10
newspaper3k==0.2.8
lxml[html_clean]==5.1.0

# Sentiment Analysis
transformers==4.37.0
torch==2.1.2
vaderSentiment==3.3.2

# Utilities
python-dotenv==1.0.0
pydantic-settings==2.1.0
httpx==0.26.0
